Move functions of mac crafter into different files
