+++ admin/osx/mac-crafter/Sources/Utils/Install.swift
@@ -1,0 +1,32 @@
+/*
+ * Copyright (C) 2024 by Claudio Cambra <claudio.cambra@nextcloud.com>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ */
+
+import Foundation
+
+func installIfMissing(
+    _ command: String,
+    _ installCommand: String,
+    installCommandEnv: [String: String]? = nil
+) {
+    if commandExists(command) {
+        print("\(command) is installed.")
+    } else {
+        print("\(command) is missing. Installing...")
+        guard shell(installCommand, env: installCommandEnv) == 0 else {
+            print("Failed to install \(command).")
+            exit(1)
+        }
+        print("\(command) installed.")
+    }
+}
